fix: pplCount와 goalPpl 타입 불일치 수정 (#64)

+++ src/main/java/server/MATE/domain/test/entity/Test.java
@@ -97,7 +97,7 @@ public class Test extends BaseEntity {
         if (this.approvalStatus != ApprovalStatus.ACCEPTED || this.testStatus != TestStatus.IN_PROGRESS) {
             throw new BaseException(BaseErrorCode.PARTICIPATION_002);
         }
-        if (this.pplCount >= this.goalPpl) {
+        if (this.pplCount >= this.goalPpl.longValue()) {
             throw new BaseException(BaseErrorCode.PARTICIPATION_004);
         }
     }
